+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/MetroDirectives.kt
@@ -22,9 +22,9 @@ object MetroDirectives : SimpleDirectivesContainer() {
     valueDirective("Enable/disable shrinking of unused bindings.") { it.toBoolean() }
   val CHUNK_FIELD_INITS by
     valueDirective("Enable/disable chunking of field initializers.") { it.toBoolean() }
-  val ENABLE_STRICT_VALIDATION by
+  val ENABLE_FULL_BINDING_GRAPH_VALIDATION by
     directive(
-      "Enable/disable strict validation of binds and provides declarations even if they are unused."
+      "Enable/disable full binding graph validation of binds and provides declarations even if they are unused."
     )
 
   // Dependency directives.