+++ compiler-tests/src/test/kotlin/dev/zacsweers/metro/compiler/MetroExtensionRegistrarConfigurator.kt
@@ -67,7 +67,7 @@ class MetroExtensionRegistrarConfigurator(testServices: TestServices) :
         chunkFieldInits =
           module.directives.singleOrZeroValue(MetroDirectives.CHUNK_FIELD_INITS)
             ?: optionDefaults.chunkFieldInits,
-        enableStrictValidation = MetroDirectives.ENABLE_STRICT_VALIDATION in module.directives,
+        enableFullBindingGraphValidation = MetroDirectives.ENABLE_FULL_BINDING_GRAPH_VALIDATION in module.directives,
         generateJvmContributionHintsInFir =
           MetroDirectives.GENERATE_JVM_CONTRIBUTION_HINTS_IN_FIR in module.directives,
         publicProviderSeverity =