+++ compiler/src/test/kotlin/dev/zacsweers/metro/compiler/ir/RenderingTest.kt
@@ -66,7 +66,7 @@ class RenderingTest : MetroCompilerTest() {
           .trimIndent(),
         extraImports = arrayOf("kotlin.reflect.KClass"),
       ),
-      options = metroOptions.copy(reportsDestination = reportsDir, enableStrictValidation = true),
+      options = metroOptions.copy(reportsDestination = reportsDir, enableFullBindingGraphValidation = true),
     ) {
       val keysFile = reportsDir.resolve("keys-populated-ExampleGraph.txt").readText()
       assertThat(keysFile)