+++ gradle-plugin/src/functionalTest/kotlin/dev/zacsweers/metro/gradle/incremental/MetroOptionOverrides.kt
@@ -2,4 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0
 package dev.zacsweers.metro.gradle.incremental
 
-data class MetroOptionOverrides(val enableStrictValidation: Boolean? = null)
+data class MetroOptionOverrides(val enableFullBindingGraphValidation: Boolean? = null)